requirements.txt
python-chess
pygame
cairosvg
streamlit
torch 